cloud_function/requirements.txt
tweepy==3.9.0
pandas==1.1.4
python-dotenv==0.15.0
gcsfs==0.7.1